Refined the project structure. Need to further organize components folder.
