+++ components/Analysis.js
