+++ components/Entry/Info.js
