+++ components/GeneralUserInput/DismissKeyboard.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import {TouchableWithoutFeedback, Keyboard} from 'react-native';
+
+export const DismissKeyboard = ({ children }) => (
+  <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
+    {children}
+  </TouchableWithoutFeedback>
+);
