+++ components/SmallChart.js
