+++ components/react-native-pure-chart/App.test.js
