+++ components/react-native-pure-chart/LICENSE
