+++ components/react-native-pure-chart/README.md
