+++ components/react-native-pure-chart/app.json
