+++ components/react-native-pure-chart/examples/App.js
