+++ components/react-native-pure-chart/examples/App.test.js
