+++ components/react-native-pure-chart/examples/app.json
