+++ components/react-native-pure-chart/examples/bar.jpeg
