+++ components/react-native-pure-chart/examples/customValueRenderer.png
