+++ components/react-native-pure-chart/examples/line.jpeg
