+++ components/react-native-pure-chart/examples/multiseries_bar.jpeg
