+++ components/react-native-pure-chart/examples/multiseries_line.jpeg
