+++ components/react-native-pure-chart/examples/package.json
