+++ components/react-native-pure-chart/examples/pie.jpg
