+++ components/react-native-pure-chart/examples/pure-chart/common.js
