+++ components/react-native-pure-chart/examples/pure-chart/components/column-chart-item.js
