+++ components/react-native-pure-chart/examples/pure-chart/components/line-chart.js
