+++ components/react-native-pure-chart/examples/pure-chart/components/pie-chart.js
