+++ components/react-native-pure-chart/examples/pure-chart/index.js
