+++ components/react-native-pure-chart/examples/sample.gif
