+++ components/react-native-pure-chart/package.json
