+++ screens/Entry.js
@@ -6,17 +6,17 @@ import {logged_in} from '../api'
 
 import {Header} from 'react-native-elements'
 import SideMenu from 'react-native-side-menu'
-import Menu from './components/Menu'
-import Info from './Info';
-import Setting from './Setting'
-import In from './In'
-import SettingSync from './SettingSync'
-import SettingGoal from './SettingGoal'
-import SettingUnits from './SettingUnits'
-import SettingTheme from './SettingTheme'
-import SettingAbout from './SettingAbout'
+import Menu from '../components/Menu'
+import Info from '../components/Entry/Info';
+import Setting from './Setting/Setting'
+import In from '../components/Entry/In'
+import SettingSync from './Setting/SettingSync'
+import SettingGoal from './Setting/SettingGoal'
+import SettingUnits from './Setting/SettingUnits'
+import SettingTheme from './Setting/SettingTheme'
+import SettingAbout from './Setting/SettingAbout'
 
-const screens={Info,Setting,In,SettingSync,SettingGoal,SettingUnits,SettingTheme,SettingAbout} //after login
+const screens={Info,Setting,In,SettingSync,SettingGoal,SettingUnits,SettingTheme,SettingAbout}
 const screenTitle={In:'Home',Info:'Info',Setting:'Setting'}
 
 export default gentle_calmm (class Entry extends Component{
@@ -53,7 +53,7 @@ export default gentle_calmm (class Entry extends Component{
 		var screenName = this.state.screen
 		var ActiveScreen = screens[screenName]
 		let screenIsIn = screenName=='In'
-		let isSettingSon = /Setting.+/.test(screenName) //assumed all setting pages are named as 'Setting.+'
+		let isSettingSon = /Setting.+/.test(screenName)
 		return (
 		<SideMenu
 		menu={<Menu screenTitle={screenTitle} selected={screenName} onItemSelected={this.onMenuItemSelected}/>}
