+++ screens/SmallChart.js
@@ -4,8 +4,8 @@ import ChartBar from './ChartBar'
 
 export default class SmallChart extends Component{
 	state={chartlayout:null,layout:null}
-	
-	
+
+
 	getBars(data,height,width){
 		let max=Math.max(...data)
 		let w=width/data.length/2
@@ -13,7 +13,7 @@ export default class SmallChart extends Component{
 		let output=data.map((d,i)=><ChartBar key={i} height={h} value={d} maxValue={max} width={w*3/2} gap={w*1/2}/>)
 		return output
 	}
-	
+
 	textBox(val,unit){
 		return(
 		<View style={styles.textBox}>
@@ -21,14 +21,14 @@ export default class SmallChart extends Component{
 			<Text style={styles.unit}>{unit}</Text>
 		</View>)
 	}
-	
+
 	render(){
 		return(
 			<View style={{...styles.container,flexDirection:this.props.portrait?'row':'column'}}>
 			<View style={{...styles.barchart,flex:(this.props.portrait?6:3)}}>
 				<Text style={styles.title}>{this.props.title}</Text>
 				<View style={styles.barContainer} onLayout={(event)=>{this.setState({layout:event.nativeEvent.layout})}}>
-					{this.state.layout?this.getBars(this.props.data,this.state.layout.height,this.props.portrait?this.state.layout.width:this.props.width):null}	
+					{this.state.layout?this.getBars(this.props.data,this.state.layout.height,this.props.portrait?this.state.layout.width:this.props.width):null}
 				</View>
 			</View>
 			{this.textBox(this.props.data[0],'Kcal.')}
